+++ src/basic/components/Wrap.js
@@ -1,7 +1,7 @@
 import { CartItems } from './CartItems';
+import { CartTotal } from './CartTotal';
 import { Header } from './Header';
 import { createAddButtonElement } from '../renders/addButton';
-import { createCartTotalElement } from '../renders/cartTotal';
 import { createProductSelectElement } from '../renders/productSelect';
 import { createStockStatusElement } from '../renders/stockStatus';
 
@@ -20,7 +20,7 @@ export const Wrap = () => {
     <div class="max-w-md mx-auto bg-white rounded-xl shadow-md overflow-hidden md:max-w-2xl p-8">
       ${Header()}
       ${CartItems({ cartItems: CART_ITEMS, productList: PRODUCT_LIST })}
-      ${createCartTotalElement().outerHTML}
+      ${CartTotal({ cartItems: CART_ITEMS })}
       ${createProductSelectElement().outerHTML}
       ${
         createAddButtonElement({
